Move TestContext up to allow use in Gradle Plugin test
